tests/wpt/tests/html/canvas/element/text/2d.text.measure.text-clusters-range.tentative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.measure.text-clusters-range.tentative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("/fonts/CanvasTest.ttf");
}
</style>
<body class="show_output">

<h1>2d.text.measure.text-clusters-range.tentative</h1>
<p class="desc">Test that getTextClusters() and fillTextCluster() correctly render different ranges of the input text.</p>


<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="400" height="300"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
promise_test(async t => {

  var canvas = document.getElementById('c');
  var ctx = canvas.getContext('2d');

  // Renders all the clusters in the list from position (x, y).
  function renderClusters(clusters, x, y) {
    for (const cluster of clusters) {
      ctx.fillTextCluster(cluster, x, y);
    }
  }

  await document.fonts.ready;

  ctx.font = '50px CanvasTest';
  ctx.textAlign = 'left';
  ctx.textBaseline = 'top';
  const text = 'EEEEE';
  let tm = ctx.measureText(text);

  // Background color.
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, canvas.width, canvas.height);

  ctx.fillStyle = '#0f0';

  // Without the first character.
  renderClusters(tm.getTextClusters(1, 5), 0, 0);
  _assertPixelApprox(canvas, 5,5, 255,0,0,255, 2);
  _assertPixelApprox(canvas, 55,5, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 105,5, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 155,5, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 205,5, 0,255,0,255, 2);
  // Without the last character.
  renderClusters(tm.getTextClusters(0, 4), 0, 100);
  _assertPixelApprox(canvas, 5,105, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 55,105, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 105,105, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 155,105, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 205,105, 255,0,0,255, 2);
  // Only the middle character.
  renderClusters(tm.getTextClusters(2, 3), 0, 150);
  _assertPixelApprox(canvas, 5,155, 255,0,0,255, 2);
  _assertPixelApprox(canvas, 55,155, 255,0,0,255, 2);
  _assertPixelApprox(canvas, 105,155, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 155,155, 255,0,0,255, 2);
  _assertPixelApprox(canvas, 205,155, 255,0,0,255, 2);

}, "Test that getTextClusters() and fillTextCluster() correctly render different ranges of the input text.");
</script>
